dubstiuted near-env with near-actix-utils and near-logger-utils
