+++ chain/chain/Cargo.toml
@@ -27,7 +27,7 @@ near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
 near-metrics = { path = "../../core/metrics" }
 near-pool = { path = "../pool" }
-near-env = {path = "../../env"}
+near-logger-utils = {path = "../../test-utils/logger"}
 
 [features]
 # if enabled, we assert in most situations that are impossible unless some byzantine behavior is observed.
