+++ chain/client/tests/query_client.rs
@@ -3,7 +3,7 @@ use futures::{future, FutureExt};
 
 use near_client::test_utils::setup_no_network;
 use near_client::Query;
-use near_env::init_test_logger;
+use near_logger_utils::init_test_logger;
 use near_primitives::types::BlockIdOrFinality;
 use near_primitives::views::{QueryRequest, QueryResponseKind};
