+++ chain/network/Cargo.toml
@@ -31,7 +31,7 @@ near-metrics = { path = "../../core/metrics" }
 near-chain = { path = "../chain" }
 
 [dev-dependencies]
-near-env = {path = "../../env"}
+near-logger-utils = {path = "../../test-utils/logger"}
 tempdir = "0.3.7"
 bencher = "0.1.5"
 num-rational = "0.2.4"
