+++ chain/network/tests/peer_handshake.rs
@@ -9,7 +9,7 @@ use actix::System;
 use futures::{future, FutureExt};
 
 use near_client::{ClientActor, ViewClientActor};
-use near_env::init_test_logger;
+use near_logger_utils::init_test_logger;
 use near_network::test_utils::{convert_boot_nodes, open_port, GetInfo, StopSignal, WaitOrTimeout};
 use near_network::types::{NetworkViewClientMessages, NetworkViewClientResponses};
 use near_network::{NetworkClientResponses, NetworkConfig, PeerManagerActor};
