+++ chain/network/tests/runner/mod.rs
@@ -13,7 +13,7 @@ use near_chain::ChainGenesis;
 use near_chain_configs::ClientConfig;
 use near_client::{ClientActor, ViewClientActor};
 use near_crypto::KeyType;
-use near_env::init_test_logger;
+use near_logger_utils::init_test_logger;
 use near_network::test_utils::{
     convert_boot_nodes, expected_routing_tables, open_port, peer_id_from_seed, BanPeerSignal,
     GetInfo, StopSignal, WaitOrTimeout,
