+++ chain/network/tests/stress_network.rs
@@ -8,7 +8,7 @@ use futures::FutureExt;
 use tracing::info;
 
 use near_client::{ClientActor, ViewClientActor};
-use near_env::init_test_logger_allow_panic;
+use near_logger_utils::init_test_logger_allow_panic;
 use near_network::test_utils::{convert_boot_nodes, open_port, GetInfo, StopSignal, WaitOrTimeout};
 use near_network::types::{NetworkViewClientMessages, NetworkViewClientResponses};
 use near_network::{NetworkClientResponses, NetworkConfig, PeerManagerActor};
