+++ neard/tests/rpc_nodes.rs
@@ -5,8 +5,8 @@ use tempdir::TempDir;
 
 use near_client::{GetBlock, TxStatus};
 use near_crypto::{InMemorySigner, KeyType};
-use near_env::init_integration_logger;
 use near_jsonrpc::client::new_client;
+use near_logger_utils::init_integration_logger;
 use near_network::test_utils::WaitOrTimeout;
 use near_primitives::serialize::to_base64;
 use near_primitives::transaction::SignedTransaction;
