+++ test-utils/logger/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "near-env"
+name = "near-logger-utils"
 version = "0.1.0"
 authors = ["Near Inc <hello@nearprotocol.com>"]
 edition = "2018"
 
 [dependencies]
-actix = "0.9.0"
+near-actix-utils = { path = "../../utils/actix" }
 tracing-subscriber = "0.2.4"
