+++ test-utils/state-viewer/src/main.rs
@@ -6,7 +6,7 @@ use ansi_term::Color::Red;
 use clap::{App, Arg, SubCommand};
 
 use near_chain::{ChainStore, ChainStoreAccess, RuntimeAdapter};
-use near_env::init_integration_logger;
+use near_logger_utils::init_integration_logger;
 use near_network::peer_store::PeerStore;
 use near_primitives::block::BlockHeader;
 use near_primitives::hash::CryptoHash;
