+++ utils/actix/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "near-actix-utils"
+version = "0.1.0"
+authors = ["Near Inc <hello@nearprotocol.com>"]
+edition = "2018"
+
+[dependencies]
+actix = "0.9.0"
